don't try to delete the backup directory if it doesn't exist
git-svn-id: svn://svn.zope.org/repos/main@25624 62d5b8a3-27da-0310-9561-8e5933582275

+++ zpkgtools/trunk/doc/mkapiref.sh
@@ -19,7 +19,7 @@ epydoc -o doc/apiref -n 'Zope Packaging Tools' --docformat restructuredtext \
 
 ERR=$?
 
-if [ $ERR -eq 0 ] ; then
+if [ $ERR -eq 0 -a -d doc/apiref-backup ] ; then
     rm -r doc/apiref-backup || exit $?
 fi
